+++ src/api/api-utils.js
@@ -1,6 +1,5 @@
 import { getConfig } from '../config/config'
 import urljoin from 'url-join'
-import { getAuthToken } from '../auth/auth'
 
 function getEndpoint (service) {
   return urljoin(getConfig().api.baseURL, service)
@@ -10,11 +9,6 @@ function requestProperties (method = 'GET') {
   let headers = {
     'Access-Control-Allow-Origin': '*',
     'Content-Type': 'application/json'
-  }
-
-  let authToken = getAuthToken()
-  if (authToken) {
-    headers['Authorization'] = authToken
   }
 
   return {