+++ src/utils/utils.js
@@ -6,3 +6,7 @@ export const promisify = (action) =>
       resolve(res)
     })
   )
+
+export const getTimestamp = () => {
+  return Date.now() / 1000
+}